# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: VPN Admin API
  description: Control plane for commercial VPN (AmneziaWG).
  version: 0.1.0-rc.1
paths:
...
  /api/v1/servers/{server_id}:
    get:
      tags:
      - servers
      summary: Get Server
      operationId: get_server_api_v1_servers__server_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags:
      - servers
      summary: Delete Server
      description: Remove a server from inventory. Fails with 409 if devices reference
        it.
      operationId: delete_server_api_v1_servers__server_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
    patch:
      tags:
      - servers
      summary: Update Server
      operationId: update_server_api_v1_servers__server_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/servers/{server_id}/cert-status:
    get:
      tags:
      - servers
      summary: Get Server Cert Status
      description: 'Return mTLS cert status for the server (from DB: fingerprint,
        expires_at).'
      operationId: get_server_cert_status_api_v1_servers__server_id__cert_status_get
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerCertStatusOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/servers/{server_id}/ips:
    get:
      tags:
      - servers
      summary: List Server Ips
      description: List IP addresses for a server.
      operationId: list_server_ips_api_v1_servers__server_id__ips_get
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerIpListOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags:
      - servers
      summary: Create Server Ip
      description: Add an IP address to a server.
      operationId: create_server_ip_api_v1_servers__server_id__ips_post
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerIpCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerIpOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/servers/{server_id}/ips/{ip_id}:
    patch:
      tags:
      - servers
      summary: Update Server Ip
      description: Update role or state of a server IP.
      operationId: update_server_ip_api_v1_servers__server_id__ips__ip_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      - name: ip_id
        in: path
        required: true
        schema:
          type: string
          title: Ip Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerIpUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerIpOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags:
      - servers
      summary: Delete Server Ip
      description: Remove an IP address from a server.
      operationId: delete_server_ip_api_v1_servers__server_id__ips__ip_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: server_id
        in: path
        required: true
        schema:
          type: string
          title: Server Id
      - name: ip_id
        in: path
        required: true
        schema:
          type: string
          title: Ip Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ServerCertStatusOut:
      properties:
        fingerprint:
          anyOf:
          - type: string
          - type: 'null'
          title: Fingerprint
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        last_rotation_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Rotation At
      type: object
      title: ServerCertStatusOut
      description: mTLS cert status for agent/server (from env or DB).
...
    ServerIpCreate:
      properties:
        ip:
          type: string
          title: Ip
          description: IPv4 or IPv6 address
        role:
          type: string
          enum:
          - primary
          - secondary
          title: Role
          default: secondary
        state:
          type: string
          enum:
          - good
          - bad
          - blocked
          - suspected_dpi
          title: State
          default: good
      type: object
      required:
      - ip
      title: ServerIpCreate
    ServerIpListOut:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ServerIpOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - total
      title: ServerIpListOut
    ServerIpOut:
      properties:
        id:
          type: string
          title: Id
        server_id:
          type: string
          title: Server Id
        ip:
          type: string
          title: Ip
        role:
          type: string
          title: Role
        state:
          type: string
          title: State
        last_checked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Checked At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - server_id
      - ip
      - role
      - state
      - created_at
      title: ServerIpOut
    ServerIpUpdate:
      properties:
        role:
          anyOf:
          - type: string
            enum:
            - primary
            - secondary
          - type: 'null'
          title: Role
        state:
          anyOf:
          - type: string
            enum:
            - good
            - bad
            - blocked
            - suspected_dpi
          - type: 'null'
          title: State
      type: object
      title: ServerIpUpdate
...
    ServerOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        region:
          type: string
          title: Region
        api_endpoint:
          type: string
          title: Api Endpoint
        vpn_endpoint:
          anyOf:
          - type: string
          - type: 'null'
          title: Vpn Endpoint
        public_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Public Key
        status:
          type: string
          enum:
          - online
          - offline
          - degraded
          - unknown
          title: Status
        is_active:
          type: boolean
          title: Is Active
        health_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Health Score
        is_draining:
          type: boolean
          title: Is Draining
          default: false
        max_connections:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Connections
        created_at:
          type: string
          format: date-time
          title: Created At
        last_seen_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Seen At
        last_snapshot_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Snapshot At
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        auto_sync_enabled:
          type: boolean
          title: Auto Sync Enabled
          default: false
        auto_sync_interval_sec:
          type: integer
          title: Auto Sync Interval Sec
          default: 60
        ops_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Ops Notes
        ops_notes_updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Ops Notes Updated At
        ops_notes_updated_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Ops Notes Updated By
        cert_fingerprint:
          anyOf:
          - type: string
          - type: 'null'
          title: Cert Fingerprint
        cert_expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Cert Expires At
      type: object
      required:
      - id
      - name
      - region
      - api_endpoint
      - vpn_endpoint
      - public_key
      - status
      - is_active
      - created_at
      title: ServerOut
...
    ServerUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        region:
          anyOf:
          - type: string
          - type: 'null'
          title: Region
        api_endpoint:
          anyOf:
          - type: string
          - type: 'null'
          title: Api Endpoint
        vpn_endpoint:
          anyOf:
          - type: string
          - type: 'null'
          title: Vpn Endpoint
        public_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Public Key
        ops_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Ops Notes
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        auto_sync_enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Auto Sync Enabled
        auto_sync_interval_sec:
          anyOf:
          - type: integer
          - type: 'null'
          title: Auto Sync Interval Sec
      type: object
      title: ServerUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ErrorInfo:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
      required:
      - code
      - message
    ErrorMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
        request_id:
          type: string
      required:
      - timestamp
      - code
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          $ref: '#/components/schemas/ErrorInfo'
        meta:
          $ref: '#/components/schemas/ErrorMeta'
      required:
      - success
      - data
      - error
      - meta
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Bot API key (X-API-Key header).
    AgentToken:
      type: apiKey
      in: header
      name: X-Agent-Token
      description: Node-agent shared token. mTLS required at reverse-proxy.
  responses:
    ErrorResponse:
      description: Unified error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'